isolate FutureBuilder to Text or Waiting
Using a FutureBuilder in a StatelessWidget kind of implies a state of sorts.
There is one state (or snapshot) prior to completing the future, and then
another state (again, snapshot) when the future completes (becomes the present state).
However, there are no redos or hot reloads. It's therefore a linear state
rather than circular or cyclical. Linear and solo-directional. But if it has
an error, then it's no longer linear, but kind of a branching state.